+++ backend/src/topic/topic.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
 import { Topic } from './topic.entity';
 import { TopicDto } from './dto/topic.dto';
 import { TopicRepository } from './topic.repository';
@@ -23,28 +23,35 @@ export class TopicService {
     }
 
     async getTopicById(id: number): Promise<Topic>{
-        const found = await this.topicRepository.findOne(id);
+        const found = await this.topicRepository.getTopicDetail(id);
         if (!found){
           throw new  NotFoundException(`Topic with ID: ${id} not found`);
         }
         return found;
     }
-    
 
     async createTopic(topicDto: TopicDto, name, req): Promise<Topic> {
       const section = await this.sectionsService.getSectionByName(name);
       return this.topicRepository.createTopic(topicDto, section.id, req.user.id);
     }
 
-    async deleteTopic(id: number): Promise<void> {
-      const result = await this.topicRepository.delete(id);
-      if(result.affected === 0){
-        throw new  NotFoundException(`Topic with ID: ${id} not found`);
-      }
+    async deleteTopic(id: number, req): Promise<void> {
+      const user = await this.usersService.findById(req.user.id);
+        if(user.admin){
+          const result = await this.topicRepository.delete(id);
+          if(result.affected === 0){
+            throw new  NotFoundException(`Topic with ID: ${id} not found`);
+          }
+        }
     }
 
-    async updateTopic(id: number, topicUpdateDto: TopicUpdateDto) {
-      return this.topicRepository.update(id, topicUpdateDto);
+    async updateTopic(id: number, topicUpdateDto: TopicUpdateDto, req) {
+      const topic = await this.topicRepository.getTopicDetail(id);
+        const user = await this.usersService.findById(req.user.id);
+        if(user.username == topic.editor.username|| user.admin){
+          return this.topicRepository.update(id, topicUpdateDto);
+        }
+        throw new BadRequestException({error: "vous ne possédez pas les droits sur ce topic"});  
     }
 
 }
